change uint8_t to uint16_t in the settings union

+++ sonoff/settings.h
@@ -154,16 +154,16 @@ typedef union {
 typedef union {
   uint16_t data;
   struct {
-    uint8_t pinmode : 3;                   // Pin mode (1 through 6)
-    uint8_t pullup : 1;                    // Enable internal weak pull-up resistor
-    uint8_t saved_state : 1;               // Save output state, if used.
-    uint8_t int_report_mode : 2;           // Interrupt reporting mode 0 = immediate telemetry & event, 1 = immediate event only, 2 = immediate telemetry only
-    uint8_t int_report_defer : 4;          // Number of interrupts to ignore until reporting (default 0, max 15)
-    uint8_t int_count_sec : 1;             // 0 = Not, 1 = Report number of interrupts since last FUNC_EVERY_SECOND callback if more than 0 occured (overrides int_report_mode if enabled)
-    uint8_t int_count_min : 1;             // 0 = Not, 1 = Report number of interrupts since last FUNC_EVERY_SECOND*60 callback if more than 0 occured (overrides int_report_mode if enabled)
-    uint8_t spare13 : 1;
-    uint8_t spare14 : 1;
-    uint8_t spare15 : 1;
+    uint16_t pinmode : 3;                   // Pin mode (1 through 6)
+    uint16_t pullup : 1;                    // Enable internal weak pull-up resistor
+    uint16_t saved_state : 1;               // Save output state, if used.
+    uint16_t int_report_mode : 2;           // Interrupt reporting mode 0 = immediate telemetry & event, 1 = immediate event only, 2 = immediate telemetry only
+    uint16_t int_report_defer : 4;          // Number of interrupts to ignore until reporting (default 0, max 15)
+    uint16_t int_count_sec : 1;             // 0 = Not, 1 = Report number of interrupts since last FUNC_EVERY_SECOND callback if more than 0 occured (overrides int_report_mode if enabled)
+    uint16_t int_count_min : 1;             // 0 = Not, 1 = Report number of interrupts since last FUNC_EVERY_SECOND*60 callback if more than 0 occured (overrides int_report_mode if enabled)
+    uint16_t spare13 : 1;
+    uint16_t spare14 : 1;
+    uint16_t spare15 : 1;
   };
 } Mcp230xxCfg;
